+++ source/Dragon.h
@@ -10,7 +10,7 @@ public:
     static double get_HP();
     static double get_power();
 private:
-    static double HP;
+    static int HP;
     static double power;
     static Dragon * p_instance;
     Dragon() = default;
